feat: enable protocol V1 on mainnet
