+++ crates/portalnet/tests/overlay.rs
@@ -10,7 +10,8 @@ use ethportal_api::{
         distance::XorMetric,
         enr::{Enr, SszEnr},
         network::Subnetwork,
-        portal_wire::{Content, Message, MAINNET},
+        network_spec::MAINNET,
+        portal_wire::{Content, Message},
     },
     utils::bytes::hex_encode_upper,
     OverlayContentKey,
